+++ packages/flutter/lib/src/material/popup_menu.dart
@@ -15,7 +15,7 @@ import 'icon_theme.dart';
 import 'icon_theme_data.dart';
 import 'icons.dart';
 import 'ink_well.dart';
-import 'list_item.dart';
+import 'list_tile.dart';
 import 'material.dart';
 import 'theme.dart';
