+++ packages/flutter/lib/src/material/two_level_list.dart
@@ -10,7 +10,7 @@ import 'icon.dart';
 import 'icon_theme.dart';
 import 'icon_theme_data.dart';
 import 'icons.dart';
-import 'list_item.dart';
+import 'list_tile.dart';
 import 'theme.dart';
 import 'theme_data.dart';
